refactor(traces): docs/, constants/, types/ — and one dead module deleted (LFE-14804)
Nikita's four reads on the branch, all correct:

  fns/types.ts        -> types/treeNode.ts + types/traceSearchListItem.ts
                         (a types dump is not a fn; one type per file now)
  shared/             -> constants/traceDownloadConfig.ts
  config/             -> constants/traceViewConfig.ts
  *.md                -> components/AdvancedJsonViewer/docs/

`shared/` held exactly one constant and `config/` one config module, so both
folded into `constants/`. The feature root is now index.ts, TracePage.tsx, and
kind folders only.

contexts/jsonExpansionUtils.ts was the fourth: not a context, so it should have
moved. Chasing its consumers showed it has none — the two files that looked like
importers import a different, identically-named module in
components/table/utils/. Its only importer was its own clienttest, so it is dead
code and it is deleted (rule 20), not relocated. 3 functions, 105 lines.

That last one is a sensor nuance worth naming: rule 20 counts file-level
orphans, and a module imported by its own test is not an orphan. Dead-in-
production and orphaned are different questions.

tsc green, lint green, Tests 482 passed | 76 skipped (558).
