+++ web/src/features/traces/components/AdvancedJsonViewer/docs/LAZY_TREE_DESIGN.md
