+++ web/src/features/traces/components/AdvancedJsonViewer/docs/lazy-react.md
