+++ web/src/features/traces/components/ObservationDetailView/ObservationDetailView.tsx
@@ -57,7 +57,7 @@ import { api } from "@/src/utils/api";
 // Extracted components
 import { ObservationDetailViewHeader } from "./components/ObservationDetailViewHeader";
 import { TraceLogView } from "../TraceLogView/TraceLogView";
-import { TRACE_VIEW_CONFIG } from "@/src/features/traces/config/trace-view-config";
+import { TRACE_VIEW_CONFIG } from "@/src/features/traces/constants/traceViewConfig";
 import { useV4Beta } from "@/src/features/events/hooks/useV4Beta";
 import {
   aggregateTraceMetrics,
