+++ web/src/features/traces/components/TraceLogView/LogViewObservationCell.tsx
@@ -8,7 +8,7 @@
 import { memo, useRef, useEffect } from "react";
 import { ItemBadge } from "@/src/components/ItemBadge";
 import { usePrefetchObservation } from "@/src/features/traces/hooks/usePrefetchObservation";
-import { TRACE_VIEW_CONFIG } from "@/src/features/traces/config/trace-view-config";
+import { TRACE_VIEW_CONFIG } from "@/src/features/traces/constants/traceViewConfig";
 import { type FlatLogItem } from "./log-view-types";
 import { formatDisplayName } from "./log-view-formatters";
