+++ web/src/features/traces/components/TraceLogView/fns/flattenChronological.clienttest.ts
@@ -1,6 +1,6 @@
 // @vitest-environment jsdom
 
-import { type TreeNode } from "@/src/features/traces/fns/types";
+import { type TreeNode } from "@/src/features/traces/types/treeNode";
 import { flattenChronological } from "@/src/features/traces/components/TraceLogView/fns/flattenChronological";
 import {
   createNode,
