+++ web/src/features/traces/components/TraceLogView/useLogViewAllObservationsIO.ts
@@ -11,7 +11,7 @@ import { useState, useCallback } from "react";
 import { useQueryClient } from "@tanstack/react-query";
 import chunk from "lodash/chunk";
 import { api } from "@/src/utils/api";
-import { TRACE_VIEW_CONFIG } from "@/src/features/traces/config/trace-view-config";
+import { TRACE_VIEW_CONFIG } from "@/src/features/traces/constants/traceViewConfig";
 import { type FlatLogItem } from "./log-view-types";
 import { formatDisplayName } from "./log-view-formatters";
