+++ web/src/features/traces/components/TraceTimeline/__tests__/timeline.fixtures.ts
@@ -5,7 +5,7 @@
  */
 
 import Decimal from "decimal.js";
-import { type TreeNode } from "@/src/features/traces/fns/types";
+import { type TreeNode } from "@/src/features/traces/types/treeNode";
 import {
   type FlatTimelineItem,
   type TimelineMetrics,
