+++ web/src/features/traces/components/TraceTimeline/timeline-flattening.ts
@@ -3,7 +3,7 @@
  * Extends base tree flattening to include timeline positioning calculations.
  */
 
-import { type TreeNode } from "../../fns/types";
+import { type TreeNode } from "../../types/treeNode";
 import { type FlatTimelineItem, type TimelineMetrics } from "./types";
 import {
   calculateTimelineOffset,
