+++ web/src/features/traces/constants/traceViewConfig.ts
@@ -1,4 +1,4 @@
-import { TRACE_DOWNLOAD_OMIT_LARGE_FIELDS_THRESHOLD } from "@/src/features/traces/shared/traceDownloadConfig";
+import { TRACE_DOWNLOAD_OMIT_LARGE_FIELDS_THRESHOLD } from "@/src/features/traces/constants/traceDownloadConfig";
 
 /**
  * Configuration constants for trace views.
