+++ web/src/features/traces/fns/traceAggregation.ts
@@ -7,7 +7,7 @@
 
 import { type ObservationReturnTypeWithMetadata } from "@/src/server/api/routers/traces";
 import { type ObservationType, isGenerationLike } from "@langfuse/shared";
-import { type TreeNode } from "./types";
+import { type TreeNode } from "../types/treeNode";
 
 export interface AggregatedTraceMetrics {
   totalCost: number | null;
