+++ web/src/features/traces/types/traceSearchListItem.ts
@@ -1,0 +1,12 @@
+import type Decimal from "decimal.js";
+import type { TreeNode } from "@/src/features/traces/types/treeNode";
+
+export interface TraceSearchListItem {
+  node: TreeNode;
+  /** Root-level total cost for heatmap scaling */
+  parentTotalCost?: Decimal;
+  /** Root-level total duration for heatmap scaling */
+  parentTotalDuration?: number;
+  /** Observation ID for navigation (undefined for TRACE nodes) */
+  observationId?: string;
+}
