+++ web/src/features/traces/types/treeNode.ts
@@ -58,12 +58,3 @@ export type TreeNode = {
  * Flattened tree node for search list and virtualized rendering.
  * Contains parent-level totals for heatmap color scaling.
  */
-export interface TraceSearchListItem {
-  node: TreeNode;
-  /** Root-level total cost for heatmap scaling */
-  parentTotalCost?: Decimal;
-  /** Root-level total duration for heatmap scaling */
-  parentTotalDuration?: number;
-  /** Observation ID for navigation (undefined for TRACE nodes) */
-  observationId?: string;
-}
